requirements.txt
flask
flask-sqlalchemy
flask-login
flask
numpy
face-recognition
face-recognition-models
Flask
beautifulsoup4
requests
opencv-python
pillow
werkzeug
